requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.4
django-crispy-forms==2.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
kombu==5.3.7
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-dateutil==2.9.0.post0
redis==5.0.3
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
pillow==10.3.0
boto3~=1.34.84